requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
emoji==2.15.0
fastapi==0.121.0
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
filetype==1.2.0
frozenlist==1.8.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.3
langchain-openai==1.0.2
langchain-qdrant==1.1.0
langchain-text-splitters==1.0.0
langdetect==1.0.9
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
nltk==3.9.2
numpy==2.3.4
olefile==0.47
openai==2.7.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==6.1.3
python-dotenv==1.2.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.15.1
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.5
sentry-sdk==2.43.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured==0.18.15
unstructured-client==0.42.3
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
webencodings==0.5.1
websockets==15.0.1
wrapt==2.0.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
